Remove default pausable behavior
Remove default pausable behavior, as it is only intended for very
specific use cases in which the caller of the lambda is able to
queue up requests using SQS or similar mechanism.  If puase is
not used in this way, then using the feature is tantamount to
causing an outage - use should be opt-in rather than the reverse.